refactor: spaceHeader 구조 변경 및 적용
