+++ frontend/src/pages/guest/imageUploadPage/ImageUploadPage.tsx
@@ -11,6 +11,7 @@ import UploadBox from '../../../components/uploadBox/UploadBox';
 import { ROUTES } from '../../../constants/routes';
 import useFileUpload from '../../../hooks/@common/useFileUpload';
 import useIntersectionObserver from '../../../hooks/@common/useIntersectionObserver';
+import useLeftTimer from '../../../hooks/@common/useLeftTimer';
 import { ScrollableBlurArea } from '../../../styles/@common/ScrollableBlurArea';
 import { theme } from '../../../styles/theme';
 import { goToTop } from '../../../utils/goToTop';
@@ -33,6 +34,9 @@ const ImageUploadPage = () => {
   const { targetRef: scrollTopTriggerRef, isIntersecting: isAtPageTop } =
     useIntersectionObserver({ isInitialInView: true });
   const navigate = useNavigate();
+  const { leftTime } = useLeftTimer({
+    targetTime: mockSpaceData.expirationDate,
+  });
 
   const handleUploadClick = async () => {
     const uploadSuccess = await handleUpload();
@@ -57,10 +61,7 @@ const ImageUploadPage = () => {
       )}
 
       <S.ScrollTopAnchor ref={scrollTopTriggerRef} />
-      <SpaceHeader
-        title={`${mockSpaceData.name}`}
-        description="클릭해서 불러올 수 있어요"
-      />
+      <SpaceHeader title={`${mockSpaceData.name}`} timer={leftTime} />
       <S.UploadContainer $hasImages={hasImages}>
         <UploadBox
           mainText={`함께한 순간을 올려주세요.${hasImages ? '' : '\n사진만 올릴 수 있습니다.'}`}
