+++ frontend/src/pages/guest/imageUploadPage/mockSpaceData.ts
@@ -1,3 +1,4 @@
 export const mockSpaceData = {
   name: '8월 버스킹',
+  expirationDate: '2025-08-06T10:20:09.761Z',
 };
