+++ frontend/src/pages/manager/spaceHome/SpaceHome.tsx
@@ -11,6 +11,7 @@ import PhotoSelectionToolBar from '../../../components/photoSelectionToolBar/Pho
 import SpaceHeader from '../../../components/spaceHeader/SpaceHeader';
 import { INFORMATION } from '../../../constants/messages';
 import useIntersectionObserver from '../../../hooks/@common/useIntersectionObserver';
+import useLeftTimer from '../../../hooks/@common/useLeftTimer';
 import useDownload from '../../../hooks/useDownload';
 import usePhotoSelect from '../../../hooks/usePhotoSelect';
 import usePhotosBySpaceCode from '../../../hooks/usePhotosBySpaceCode';
@@ -30,6 +31,9 @@ const SpaceHome = () => {
     isIntersecting: isFetchSectionVisible,
     reObserve,
   } = useIntersectionObserver({ rootMargin: '200px' });
+  const { leftTime } = useLeftTimer({
+    targetTime: mockSpaceData.expirationDate,
+  });
 
   const {
     photosList,
@@ -72,7 +76,7 @@ const SpaceHome = () => {
       <S.InfoContainer ref={scrollTopTriggerRef}>
         <SpaceHeader
           title={mockSpaceData.name}
-          description={mockSpaceData.startDate}
+          timer={leftTime}
           icon={
             <SettingSvg
               fill={theme.colors.primary20}
