+++ frontend/src/pages/manager/spaceHome/mockSpaceData.ts
@@ -4,4 +4,5 @@ export const mockSpaceData = {
   participantsCount: 80,
   photosCount: 6,
   code: '1234567890',
+  expirationDate: '2025-08-06T03:20:09.761Z',
 };
