+++ frontend/src/stories/SpaceHeader.stories.tsx
@@ -13,7 +13,25 @@ type Story = StoryObj<typeof meta>;
 export const Default: Story = {
   args: {
     title: '제목임',
-    description: '설명임',
+    timer: {
+      days: 0,
+      hours: 1,
+      minutes: 45,
+      seconds: 30,
+    },
+    icon: <SaveIcon />,
+  },
+};
+
+export const Urgent: Story = {
+  args: {
+    title: '제목임',
+    timer: {
+      days: 0,
+      hours: 0,
+      minutes: 45,
+      seconds: 30,
+    },
     icon: <SaveIcon />,
   },
 };
